✨feat(boards): 컴포넌트 분리 및 반응형 작업 완료
