+++ src/app/boards/components/Button.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { PropsWithChildren } from "react";
 
 interface IButtonProps extends PropsWithChildren {
