+++ src/app/boards/components/Dropdown.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 import React, { useState } from "react";
