+++ src/app/boards/components/bestArticles/BestFlex.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import BestPostCard from "./BestPostCard";
