+++ src/app/boards/components/bestArticles/BestPostList.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import useWindowSize from "@/hooks/useWindowSize";
 import React from "react";
 import BestSwiper from "./BestSwiper";
@@ -5,12 +7,8 @@ import BestFlex from "./BestFlex";
 
 function BestPostList() {
   const windowSize = useWindowSize();
-  const isMobile = windowSize < 640;
+  const isMobile = windowSize < 640 && windowSize !== 0;
 
-  if (windowSize === 0) {
-    return;
-  }
-  // eslint-disable-next-line consistent-return
   return <div className="w-full overflow-hidden">{isMobile ? <BestSwiper /> : <BestFlex />}</div>;
 }
