+++ src/app/boards/components/search/SearchForm.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import SearchInput from "./SearchInput";
 import Button from "../Button";
