+++ src/app/boards/components/search/SearchInput.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React from "react";
 import SearchIcon from "@/assets/icons/search.svg";
